Implementacja endpointu odpowiedzialnego za usuwanie aut. Implementacja testów dla kontrolera powiązanego z wynajmem. Ujednolicenie typów w modelach. Uproszczenie inicjalizacji. Drobne zwiększenie czytelności kodu.

+++ src/main/java/com/example/carrentalsystem/Models/CarImage.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 @Setter
 public class CarImage {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer imageID;
+    private Long imageID;
 
     @Lob
     @Column(nullable = false)
