+++ src/main/java/com/example/carrentalsystem/Repositories/CarImageRepository.java
@@ -3,6 +3,6 @@ package com.example.carrentalsystem.Repositories;
 import com.example.carrentalsystem.Models.CarImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarImageRepository extends JpaRepository<CarImage, Integer> {
-    CarImage getByImageID(Integer imageID);
+public interface CarImageRepository extends JpaRepository<CarImage, Long> {
+    CarImage getByImageID(Long imageID);
 }
